requirements.txt
alembic==1.9.1
anyio==3.6.2
asttokens==2.2.1
asyncpg==0.27.0
attrs==22.1.0
backcall==0.2.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
coverage==7.0.5
crashtest==0.4.1
cryptography==37.0.4
decorator==5.1.1
distlib==0.3.6
dulwich==0.20.46
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.89.1
filelock==3.8.0
flake8==6.0.0
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipython==8.8.0
isort==5.11.4
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.14.0
msgpack==1.0.4
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
Pygments==2.14.0
pyrsistent==0.18.1
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
python-dotenv==0.21.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
stack-data==0.6.2
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.5
traitlets==5.8.1
trove-classifiers==2022.12.22
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1